Load more spaces
This features allows to load more than the 20 first spaces of the user, in the share extension.
When the spaces are loaded successfully, the app checks if there are more to load or not.
If yes, when the user reaches the bottom, the app loads the next 20 spaces, and so on.