MOB-2155 : make Authentication for REST API work with HTTPS
The cookie for authentication for REST calls (JSESSIONID) is not used when connecting on an HTTPS eXo instance. This is because this cookie is set as Secure and calling CookieManager.getCookie with the domain as an argument returns only the non-secure cookies. This method must be called with the "https://" prefix to get all cookies.
This fix uses the full url to be sure to retrieve all cookies.