MOB-2170 : do not reuse intents when creating notifications
The first parameter of the getPendingIntent method refers is the id of the intent. If an intent with the same id already uses, it will reuse it.
So when multiple notifications were sent, the previous notifications were used and their extra parameters were overridden.
This fix sets an unique id for each notification pending intent (by reusing the notification id) to make sure extra parameters are not overridden.