MOB-2197 : handle file download (#10)
The file download event was not handled, which resulted in no action when a file download was requested.

This fix adds the fle download handler.
Also since Android API 23 (M), the permissions must be acknowledged at runtime (when used) instead of during the app installation. So this fix check the write storage permission when a file download is requested.