MOB-2197 : rely on filename extension instead of mimetype for files downloads
When a file is downloaded, the name used to save the file on the device is calculated from the contentDisposition and the mimetype passed in the URL.
When a file is downoaded from an activity or from the preview, the download URL does not contain the filename (for example /portal/download?resourceId=903576950), the contentDisposition contains the real filename (for example "attachment;filename="my-file.pdf") and the mimetype is set to "application/zip" (for some reasons...).
The method used to calculate the filename is android.webkit.URLUtil#guessFileName.
If the mimeType and the contentDisposition are different, this method replaces the extension of the file by the extension of the mimetype (so in our example it would replace ".pdf" by ".zip").
That's why it stores a zip file instead of the original file type.
This issue does not occurs when the download URL ends with the filename, like in the Documents application (for example "rest/contents/download/collaboration/Users/r___/ro___/roo___/root/Public/Activity+Stream+Documents/my-file.pdf").

This fix set the mimeType to null when calculating the filename in order to ignore it and always rely on the URL or the contentDisposition.