TASK-57263: Push notifications still received after logout
Prior to this change, clearWebViewData was clearing all app data before logout includes the user context information those are needed in rest filter verifications to allow to proceed the call of the rest endpoint to delete the user token.
This PR should make sure to delete user token before clearing the data.