fix: display correctly uploaded images on chat - EXO-70047 (#165)
Chat images were displayed without the Done button, thus we can not get back to the chat page once done. Besides, images are rendered bigger than the screen and it was not possible to zoom into them.
The fix gets correctly the content type to detect when an image is loaded, then it displays the Done button and enables the zoom in/out

After opening an image shared in a chat message, the back button and the Done button should close the new webview and display the original webview to avoid loosing the initial context .
The fix displays correctly the image following the device size, and provide the same behaviour when clicking on Done or pressing back button.

Files having HTML special chars in their names are not recognized by the Mimetype resolver.
the fix decodes their names to get the original name and extract the extension correctly.
Besides this fix re-enables downloading directly files from the Chat